refactor: 폴더명, 클래스명 변경
- Books -> Book
- api 주소만 books 사용
