+++ src/app.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { BooksModule } from './books/books.module';
+import { BooksModule } from './book/book.module';
 import { typeORMConfig } from './configs/typeorm.config';
 
 @Module({
