+++ src/book/book.controller.ts
@@ -1,11 +1,11 @@
 import { Body, Controller, Get, Post, Query, UsePipes, ValidationPipe } from '@nestjs/common';
 import { Book } from './book.entity';
-import { BooksService } from './books.service';
+import { BookService } from './book.service';
 import { CreateBookDto } from './dto/create-book.dto';
 
 @Controller('books')
-export class BooksController {
-  constructor(private booksService: BooksService) {}
+export class BookController {
+  constructor(private booksService: BookService) {}
   
   @Post()
   @UsePipes(ValidationPipe)
