+++ src/book/book.module.ts
@@ -1,16 +1,16 @@
 import { HttpModule } from '@nestjs/axios';
 import { Module } from '@nestjs/common';
 import { TypeOrmExModule } from 'src/db/typeorm-ex.module';
+import { BookController } from './book.controller';
 import { BookRepository } from './book.repository';
-import { BooksController } from './books.controller';
-import { BooksService } from './books.service';
+import { BookService } from './book.service';
 
 @Module({
   imports: [
     TypeOrmExModule.forCustomRepository([BookRepository]),
     HttpModule
   ],
-  controllers: [BooksController],
-  providers: [BooksService]
+  controllers: [BookController],
+  providers: [BookService]
 })
 export class BooksModule {}
